% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{absolutePanel}
\alias{absolutePanel}
\title{Create an absolutely positioned panel}
\usage{
absolutePanel(
  ...,
  top = NULL,
  left = NULL,
  right = NULL,
  bottom = NULL,
  width = NULL,
  height = NULL,
  draggable = FALSE,
  fixed = FALSE,
  cursor = "auto"
)
}
\arguments{
\item{...}{UI elements}

\item{top}{Distance from top}

\item{left}{Distance from left}

\item{right}{Distance from right}

\item{bottom}{Distance from bottom}

\item{width}{Panel width}

\item{height}{Panel height}

\item{draggable}{Allow dragging?}

\item{fixed}{Use fixed positioning?}

\item{cursor}{Cursor style}
}
\value{
An absolute panel tag
}
\description{
Create an absolutely positioned panel
}
