% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{actionButton}
\alias{actionButton}
\title{Create an action button}
\usage{
actionButton(inputId, label, icon = NULL, width = NULL, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Button label}

\item{icon}{Optional icon (from icon())}

\item{width}{Width of the button}

\item{...}{Additional attributes}
}
\value{
An action button tag
}
\description{
Create an action button
}
