% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{actionLink}
\alias{actionLink}
\title{Create an action link}
\usage{
actionLink(inputId, label, icon = NULL, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Link label}

\item{icon}{Optional icon}

\item{...}{Additional attributes}
}
\value{
An action link tag
}
\description{
Create an action link
}
