% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{animationOptions}
\alias{animationOptions}
\title{Animation options for slider input}
\usage{
animationOptions(
  interval = 1000,
  loop = FALSE,
  playButton = NULL,
  pauseButton = NULL
)
}
\arguments{
\item{interval}{Time between animation frames (ms)}

\item{loop}{Loop the animation?}

\item{playButton}{Custom play button}

\item{pauseButton}{Custom pause button}
}
\value{
Animation options list
}
\description{
Animation options for slider input
}
