% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Create a hotShiny application}
\usage{
app(ui, server, ...)
}
\arguments{
\item{ui}{UI function or object}

\item{server}{Server function}

\item{...}{Additional arguments}
}
\value{
App object
}
\description{
Create a hotShiny application
}
