% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-dynamic.R
\name{appendTab}
\alias{appendTab}
\title{Append a tab to a tabset panel}
\usage{
appendTab(inputId, tab, select = FALSE, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Tabset panel ID}

\item{tab}{Tab panel to append}

\item{select}{Select the new tab?}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Append a tab to a tabset panel
}
