% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{as.character.shiny.tag.list}
\alias{as.character.shiny.tag.list}
\title{Convert tag list to character}
\usage{
\method{as.character}{shiny.tag.list}(x, ...)
}
\arguments{
\item{x}{Tag list object}

\item{...}{Additional arguments}
}
\description{
Convert tag list to character
}
