% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{as.character.shiny.tag}
\alias{as.character.shiny.tag}
\title{Convert tag to character (HTML string)}
\usage{
\method{as.character}{shiny.tag}(x, ...)
}
\arguments{
\item{x}{Tag object}

\item{...}{Additional arguments}
}
\description{
Convert tag to character (HTML string)
}
