% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-dependency-tracker.R
\name{ast_to_expr}
\alias{ast_to_expr}
\title{Reconstruct expression from AST}
\usage{
ast_to_expr(ast)
}
\arguments{
\item{ast}{AST representation}
}
\value{
R expression
}
\description{
Reconstruct expression from AST
}
