% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{basicPage}
\alias{basicPage}
\title{Basic page (alias for bootstrapPage)}
\usage{
basicPage(..., title = NULL, theme = NULL)
}
\arguments{
\item{...}{UI elements}

\item{title}{Page title}

\item{theme}{Optional theme}
}
\value{
A basic page tag
}
\description{
Basic page (alias for bootstrapPage)
}
