% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{blockquote}
\alias{blockquote}
\title{Create a blockquote tag}
\usage{
blockquote(...)
}
\arguments{
\item{...}{Tag contents and attributes}
}
\value{
A blockquote tag
}
\description{
Create a blockquote tag
}
