% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{bootstrapPage}
\alias{bootstrapPage}
\title{Create a Bootstrap page}
\usage{
bootstrapPage(..., title = NULL, theme = NULL)
}
\arguments{
\item{...}{UI elements}

\item{title}{Page title}

\item{theme}{Optional theme}
}
\value{
A bootstrap page tag
}
\description{
Minimal Bootstrap page without extra structure.
}
