% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-graph-builder.R
\name{build_reactive_graph}
\alias{build_reactive_graph}
\title{Build reactive graph from server function}
\usage{
build_reactive_graph(server_func, source_info = NULL)
}
\arguments{
\item{server_func}{Server function}

\item{source_info}{Source file information}
}
\value{
ReactiveGraph
}
\description{
Build reactive graph from server function
}
