% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-strict-mode.R
\name{check_strict}
\alias{check_strict}
\title{Check expression in strict mode}
\usage{
check_strict(expr, app, context = NULL)
}
\arguments{
\item{expr}{Expression to check}

\item{app}{HotShinyApp instance}

\item{context}{Context information}
}
\value{
Check result
}
\description{
Check expression in strict mode
}
