% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{checkboxGroupInput}
\alias{checkboxGroupInput}
\title{Create a checkbox group input}
\usage{
checkboxGroupInput(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  width = NULL,
  choiceNames = NULL,
  choiceValues = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{choices}{List of values to select from}

\item{selected}{Initially selected value(s)}

\item{inline}{Display choices inline?}

\item{width}{Width of the container}

\item{choiceNames}{Names to display (alternative to choices)}

\item{choiceValues}{Values to return (alternative to choices)}
}
\value{
A checkbox group tag
}
\description{
Create a checkbox group input
}
