% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{checkboxInput}
\alias{checkboxInput}
\title{Create a single checkbox input}
\usage{
checkboxInput(inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{value}{Initial value (TRUE or FALSE)}

\item{width}{Width of the container}
}
\value{
A checkbox input tag
}
\description{
Create a single checkbox input
}
