% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{column}
\alias{column}
\title{Create a column within a fluid row}
\usage{
column(width, ..., offset = 0)
}
\arguments{
\item{width}{Column width (1-12 in Bootstrap grid)}

\item{...}{UI elements to include in the column}

\item{offset}{Number of columns to offset}
}
\value{
A column tag
}
\description{
Create a column within a fluid row
}
