% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-serializer.R
\name{compute_graph_hash}
\alias{compute_graph_hash}
\title{Compute graph hash for versioning}
\usage{
compute_graph_hash(graph)
}
\arguments{
\item{graph}{ReactiveGraph object}
}
\value{
Hash string
}
\description{
Compute graph hash for versioning
}
