% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{compute_node_hash}
\alias{compute_node_hash}
\title{Compute node content hash}
\usage{
compute_node_hash(node)
}
\arguments{
\item{node}{ReactiveNode}
}
\value{
Hash string
}
\description{
Compute node content hash
}
