% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{conditionalPanel}
\alias{conditionalPanel}
\title{Create a conditional panel}
\usage{
conditionalPanel(condition, ..., ns = NULL)
}
\arguments{
\item{condition}{JavaScript expression that evaluates to TRUE/FALSE}

\item{...}{UI elements}

\item{ns}{Namespace function for module}
}
\value{
A conditional panel tag
}
\description{
A panel that shows/hides based on a JavaScript condition.
}
