% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-file-watcher.R
\name{create_file_watcher}
\alias{create_file_watcher}
\title{Create a file watcher}
\usage{
create_file_watcher(check_interval = 0.5)
}
\arguments{
\item{check_interval}{Interval between checks (seconds)}
}
\value{
FileWatcher instance
}
\description{
Create a file watcher
}
