% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-message-handler.R
\name{create_message_handler}
\alias{create_message_handler}
\title{Create message handler}
\usage{
create_message_handler(app)
}
\arguments{
\item{app}{HotShinyApp instance}
}
\value{
MessageHandler instance
}
\description{
Create message handler
}
