% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-session-manager.R
\name{create_session_manager}
\alias{create_session_manager}
\title{Create session manager}
\usage{
create_session_manager()
}
\value{
SessionManager instance
}
\description{
Create session manager
}
