% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-websocket.R
\name{create_websocket_server}
\alias{create_websocket_server}
\title{Create WebSocket server for app}
\usage{
create_websocket_server(app)
}
\arguments{
\item{app}{HotShinyApp instance}
}
\value{
WebSocketServer instance
}
\description{
Create WebSocket server for app
}
