% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{createRenderFunction}
\alias{createRenderFunction}
\title{Create a custom render function}
\usage{
createRenderFunction(
  func,
  outputFunc = NULL,
  outputArgs = list(),
  cacheHint = list(),
  cacheWriteHook = NULL,
  cacheReadHook = NULL
)
}
\arguments{
\item{func}{Render function}

\item{outputFunc}{Output function (for Shiny module support)}

\item{outputArgs}{Output arguments}

\item{cacheHint}{Cache hint}

\item{cacheWriteHook}{Cache write hook}

\item{cacheReadHook}{Cache read hook}
}
\value{
Custom render function
}
\description{
Create a custom render function
}
