% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{dataTableOutput}
\alias{dataTableOutput}
\title{Create a DataTables output element}
\usage{
dataTableOutput(outputId)
}
\arguments{
\item{outputId}{Output variable for the DataTable}
}
\value{
A DataTable output tag
}
\description{
Create a DataTables output element
}
