% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{dateInput}
\alias{dateInput}
\title{Create a date input}
\usage{
dateInput(
  inputId,
  label,
  value = NULL,
  min = NULL,
  max = NULL,
  format = "yyyy-mm-dd",
  startview = "month",
  weekstart = 0,
  language = "en",
  width = NULL,
  autoclose = TRUE,
  datesdisabled = NULL,
  daysofweekdisabled = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{value}{Initial value (Date or character)}

\item{min}{Minimum selectable date}

\item{max}{Maximum selectable date}

\item{format}{Date format string}

\item{startview}{Starting view: "month", "year", "decade"}

\item{weekstart}{Day of week to start (0 = Sunday)}

\item{language}{Language code}

\item{width}{Width of the input}

\item{autoclose}{Close picker on selection?}

\item{datesdisabled}{Vector of disabled dates}

\item{daysofweekdisabled}{Vector of disabled days (0-6)}
}
\value{
A date input tag
}
\description{
Create a date input
}
