% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{dateRangeInput}
\alias{dateRangeInput}
\title{Create a date range input}
\usage{
dateRangeInput(
  inputId,
  label,
  start = NULL,
  end = NULL,
  min = NULL,
  max = NULL,
  format = "yyyy-mm-dd",
  startview = "month",
  weekstart = 0,
  language = "en",
  separator = " to ",
  width = NULL,
  autoclose = TRUE
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{start}{Initial start date}

\item{end}{Initial end date}

\item{min}{Minimum selectable date}

\item{max}{Maximum selectable date}

\item{format}{Date format string}

\item{startview}{Starting view: "month", "year", "decade"}

\item{weekstart}{Day of week to start (0 = Sunday)}

\item{language}{Language code}

\item{separator}{Separator between dates}

\item{width}{Width of the input}

\item{autoclose}{Close picker on selection?}
}
\value{
A date range input tag
}
\description{
Create a date range input
}
