% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-serializer.R
\name{deserialize_graph}
\alias{deserialize_graph}
\title{Deserialize graph from JSON}
\usage{
deserialize_graph(json)
}
\arguments{
\item{json}{JSON string}
}
\value{
ReactiveGraph object
}
\description{
Deserialize graph from JSON
}
