% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{diff_graphs}
\alias{diff_graphs}
\title{Diff two reactive graphs}
\usage{
diff_graphs(old_graph, new_graph)
}
\arguments{
\item{old_graph}{Old graph version}

\item{new_graph}{New graph version}
}
\value{
GraphDiff object
}
\description{
Diff two reactive graphs
}
