% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-strict-mode.R
\name{disable_strict_mode}
\alias{disable_strict_mode}
\title{Disable strict mode for an app}
\usage{
disable_strict_mode(app)
}
\arguments{
\item{app}{HotShinyApp instance}
}
\value{
App with strict mode disabled
}
\description{
Disable strict mode for an app
}
