% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-time-travel.R
\name{disable_time_travel}
\alias{disable_time_travel}
\title{Disable time-travel debugging}
\usage{
disable_time_travel(app)
}
\arguments{
\item{app}{HotShinyApp instance}
}
\value{
App with debugging disabled
}
\description{
Disable time-travel debugging
}
