% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{downloadButton}
\alias{downloadButton}
\title{Create a download button}
\usage{
downloadButton(outputId, label = "Download", class = NULL, icon = NULL, ...)
}
\arguments{
\item{outputId}{Output variable bound to downloadHandler}

\item{label}{Button label}

\item{class}{CSS class for button}

\item{icon}{Optional icon}

\item{...}{Additional attributes}
}
\value{
A download button tag
}
\description{
Create a download button
}
