% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{downloadHandler}
\alias{downloadHandler}
\title{Create a download handler}
\usage{
downloadHandler(filename, content, contentType = NA, outputArgs = list())
}
\arguments{
\item{filename}{Function or string for filename}

\item{content}{Function that writes content to a file}

\item{contentType}{MIME type (or function returning MIME type)}

\item{outputArgs}{Output arguments}
}
\value{
Download handler object
}
\description{
Creates a handler for file downloads.
}
