% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{downloadLink}
\alias{downloadLink}
\title{Create a download link}
\usage{
downloadLink(outputId, label = "Download", class = NULL, icon = NULL, ...)
}
\arguments{
\item{outputId}{Output variable bound to downloadHandler}

\item{label}{Link label}

\item{class}{CSS class for link}

\item{icon}{Optional icon}

\item{...}{Additional attributes}
}
\value{
A download link tag
}
\description{
Create a download link
}
