% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{dropNulls}
\alias{dropNulls}
\title{Remove NULL values from a list}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{List to filter}
}
\value{
List with NULL values removed
}
\description{
Remove NULL values from a list
}
