% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-reload-engine.R
\name{enable_hot_reload}
\alias{enable_hot_reload}
\title{Enable hot reload for an app}
\usage{
enable_hot_reload(app, watch_paths = NULL)
}
\arguments{
\item{app}{HotShinyApp instance}

\item{watch_paths}{Paths to watch (defaults to current directory)}
}
\description{
Enable hot reload for an app
}
