% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-strict-mode.R
\name{enable_strict_mode}
\alias{enable_strict_mode}
\title{Enable strict mode for an app}
\usage{
enable_strict_mode(app)
}
\arguments{
\item{app}{HotShinyApp instance}
}
\value{
App with strict mode enabled
}
\description{
Enable strict mode for an app
}
