% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-time-travel.R
\name{enable_time_travel}
\alias{enable_time_travel}
\title{Enable time-travel debugging}
\usage{
enable_time_travel(app, max_history = 1000)
}
\arguments{
\item{app}{HotShinyApp instance}

\item{max_history}{Maximum history size}
}
\value{
App with debugging enabled
}
\description{
Enable time-travel debugging
}
