% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{eventsource}
\alias{eventsource}
\title{Create a eventsource tag}
\usage{
eventsource(...)
}
\arguments{
\item{...}{Tag contents and attributes}
}
\value{
A eventsource tag
}
\description{
Create a eventsource tag
}
