% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-dependency-tracker.R
\name{expr_to_ast}
\alias{expr_to_ast}
\title{Convert expression to AST representation (serializable)}
\usage{
expr_to_ast(expr)
}
\arguments{
\item{expr}{R expression}
}
\value{
List representation of AST
}
\description{
Convert expression to AST representation (serializable)
}
