% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{figcaption}
\alias{figcaption}
\title{Create a figcaption tag}
\usage{
figcaption(...)
}
\arguments{
\item{...}{Tag contents and attributes}
}
\value{
A figcaption tag
}
\description{
Create a figcaption tag
}
