% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{fileInput}
\alias{fileInput}
\title{Create a file input}
\usage{
fileInput(
  inputId,
  label,
  multiple = FALSE,
  accept = NULL,
  width = NULL,
  buttonLabel = "Browse...",
  placeholder = "No file selected",
  capture = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{multiple}{Allow multiple file selection?}

\item{accept}{MIME types or file extensions to accept}

\item{width}{Width of the input}

\item{buttonLabel}{Label for the browse button}

\item{placeholder}{Placeholder text for file name}

\item{capture}{For mobile: "user" for front camera, "environment" for rear}
}
\value{
A file input tag
}
\description{
Create a file input
}
