% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-file-watcher.R
\name{FileWatcher}
\alias{FileWatcher}
\title{File Watcher}
\description{
Watches files for changes and triggers callbacks
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileWatcher-new}{\code{FileWatcher$new()}}
\item \href{#method-FileWatcher-watch}{\code{FileWatcher$watch()}}
\item \href{#method-FileWatcher-watch_directory}{\code{FileWatcher$watch_directory()}}
\item \href{#method-FileWatcher-start}{\code{FileWatcher$start()}}
\item \href{#method-FileWatcher-stop}{\code{FileWatcher$stop()}}
\item \href{#method-FileWatcher-check_loop}{\code{FileWatcher$check_loop()}}
\item \href{#method-FileWatcher-get_watched_files}{\code{FileWatcher$get_watched_files()}}
\item \href{#method-FileWatcher-unwatch}{\code{FileWatcher$unwatch()}}
\item \href{#method-FileWatcher-clone}{\code{FileWatcher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$new(check_interval = 0.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-watch"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-watch}{}}}
\subsection{Method \code{watch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$watch(file_path, callback = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-watch_directory"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-watch_directory}{}}}
\subsection{Method \code{watch_directory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$watch_directory(
  dir_path,
  pattern = "\\\\.R$",
  callback = NULL,
  recursive = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-start"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-stop"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-stop}{}}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-check_loop"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-check_loop}{}}}
\subsection{Method \code{check_loop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$check_loop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-get_watched_files"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-get_watched_files}{}}}
\subsection{Method \code{get_watched_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$get_watched_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-unwatch"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-unwatch}{}}}
\subsection{Method \code{unwatch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$unwatch(file_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileWatcher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileWatcher-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileWatcher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
