% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fillCol}
\alias{fillCol}
\title{Create a flexbox-based column}
\usage{
fillCol(..., flex = 1, width = "100\%")
}
\arguments{
\item{...}{UI elements}

\item{flex}{Flex values for children}

\item{width}{Column width}
}
\value{
A fill column tag
}
\description{
Create a flexbox-based column
}
