% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fillPage}
\alias{fillPage}
\title{Create a page that fills the window}
\usage{
fillPage(..., padding = 0, title = NULL, theme = NULL)
}
\arguments{
\item{...}{UI elements}

\item{padding}{Padding around the page}

\item{title}{Page title}

\item{theme}{Optional theme}
}
\value{
A fill page tag
}
\description{
Create a page that fills the window
}
