% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fillRow}
\alias{fillRow}
\title{Create a flexbox-based row}
\usage{
fillRow(..., flex = 1, height = "100\%")
}
\arguments{
\item{...}{UI elements}

\item{flex}{Flex values for children}

\item{height}{Row height}
}
\value{
A fill row tag
}
\description{
Create a flexbox-based row
}
