% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fixedPage}
\alias{fixedPage}
\title{Create a page with fixed layout}
\usage{
fixedPage(..., title = NULL, theme = NULL, lang = NULL)
}
\arguments{
\item{...}{UI elements to include in the page}

\item{title}{Page title}

\item{theme}{Optional theme CSS file}

\item{lang}{Language attribute}
}
\value{
A fixed page tag
}
\description{
Creates a Bootstrap container page with fixed width.
}
