% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fixedPanel}
\alias{fixedPanel}
\title{Create a fixed position panel}
\usage{
fixedPanel(
  ...,
  top = NULL,
  left = NULL,
  right = NULL,
  bottom = NULL,
  width = NULL,
  height = NULL,
  draggable = FALSE,
  cursor = "auto"
)
}
\arguments{
\item{...}{UI elements}

\item{top}{Distance from top}

\item{left}{Distance from left}

\item{right}{Distance from right}

\item{bottom}{Distance from bottom}

\item{width}{Panel width}

\item{height}{Panel height}

\item{draggable}{Allow dragging?}

\item{cursor}{Cursor style}
}
\value{
A fixed panel tag
}
\description{
Create a fixed position panel
}
