% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fixedRow}
\alias{fixedRow}
\title{Create a row in a fixed layout}
\usage{
fixedRow(...)
}
\arguments{
\item{...}{UI elements (typically column() calls)}
}
\value{
A fixed row tag
}
\description{
Create a row in a fixed layout
}
