% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{flowLayout}
\alias{flowLayout}
\title{Create a flow layout}
\usage{
flowLayout(..., cellArgs = list())
}
\arguments{
\item{...}{UI elements}

\item{cellArgs}{Additional arguments for each cell}
}
\value{
A flow layout tag
}
\description{
Elements flow left-to-right, wrapping to new lines.
}
