% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{fluidRow}
\alias{fluidRow}
\title{Create a row in a fluid layout}
\usage{
fluidRow(...)
}
\arguments{
\item{...}{UI elements (typically column() calls)}
}
\value{
A fluid row tag
}
\description{
Create a row in a fluid layout
}
