% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{get_dependency_nodes}
\alias{get_dependency_nodes}
\title{Get nodes that a node depends on}
\usage{
get_dependency_nodes(graph, node_id)
}
\arguments{
\item{graph}{ReactiveGraph}

\item{node_id}{Node ID}
}
\value{
List of dependency node IDs
}
\description{
Get nodes that a node depends on
}
