% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{get_dependent_nodes}
\alias{get_dependent_nodes}
\title{Get nodes that depend on a given node}
\usage{
get_dependent_nodes(graph, node_id)
}
\arguments{
\item{graph}{ReactiveGraph}

\item{node_id}{Node ID}
}
\value{
List of dependent node IDs
}
\description{
Get nodes that depend on a given node
}
