% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-reactive.R
\name{get_source_location}
\alias{get_source_location}
\title{Get source location from call stack}
\usage{
get_source_location()
}
\value{
List with file and line, or NULL
}
\description{
Get source location from call stack
}
