% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{getDefaultReactiveDomain}
\alias{getDefaultReactiveDomain}
\title{Get default reactive domain (session)}
\usage{
getDefaultReactiveDomain()
}
\value{
Current session or NULL
}
\description{
Get default reactive domain (session)
}
