% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{getQueryString}
\alias{getQueryString}
\title{Get the query string from the URL}
\usage{
getQueryString(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session object}
}
\value{
Named list of query parameters
}
\description{
Get the query string from the URL
}
