% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{getUrlHash}
\alias{getUrlHash}
\title{Get the URL hash}
\usage{
getUrlHash(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session object}
}
\value{
Hash string (without #)
}
\description{
Get the URL hash
}
