% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-navigation.R, R/ui-updates.R, R/ui-dynamic.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y

x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Default value}
}
\value{
x if not NULL, otherwise y

x if not NULL, otherwise y

x if not NULL, otherwise y
}
\description{
Null-coalescing operator

Null-coalescing operator

Null-coalescing operator
}
