% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{graph_structure_changed}
\alias{graph_structure_changed}
\title{Check if graph structure changed}
\usage{
graph_structure_changed(old_graph, new_graph)
}
\arguments{
\item{old_graph}{Old graph}

\item{new_graph}{New graph}
}
\value{
TRUE if structure changed
}
\description{
Check if graph structure changed
}
