% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-graph-builder.R
\name{GraphBuilder}
\alias{GraphBuilder}
\title{Graph Builder}
\description{
Builds reactive graph from Shiny code
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GraphBuilder-new}{\code{GraphBuilder$new()}}
\item \href{#method-GraphBuilder-build_from_server}{\code{GraphBuilder$build_from_server()}}
\item \href{#method-GraphBuilder-register_input}{\code{GraphBuilder$register_input()}}
\item \href{#method-GraphBuilder-register_reactive}{\code{GraphBuilder$register_reactive()}}
\item \href{#method-GraphBuilder-register_observer}{\code{GraphBuilder$register_observer()}}
\item \href{#method-GraphBuilder-register_render}{\code{GraphBuilder$register_render()}}
\item \href{#method-GraphBuilder-get_graph}{\code{GraphBuilder$get_graph()}}
\item \href{#method-GraphBuilder-clone}{\code{GraphBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-build_from_server"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-build_from_server}{}}}
\subsection{Method \code{build_from_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$build_from_server(server_func, source_info = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-register_input"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-register_input}{}}}
\subsection{Method \code{register_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$register_input(input_name, source = NULL, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-register_reactive"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-register_reactive}{}}}
\subsection{Method \code{register_reactive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$register_reactive(
  expr,
  name = NULL,
  deps = NULL,
  source = NULL,
  env = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-register_observer"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-register_observer}{}}}
\subsection{Method \code{register_observer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$register_observer(
  expr,
  deps = NULL,
  priority = 0L,
  once = FALSE,
  suspended = FALSE,
  source = NULL,
  env = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-register_render"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-register_render}{}}}
\subsection{Method \code{register_render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$register_render(
  render_type,
  output_name,
  expr,
  deps = NULL,
  source = NULL,
  env = NULL,
  outputArgs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$get_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GraphBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
