% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{GraphDiff}
\alias{GraphDiff}
\title{Graph Diff Result}
\description{
Result of comparing two graphs
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GraphDiff-new}{\code{GraphDiff$new()}}
\item \href{#method-GraphDiff-has_changes}{\code{GraphDiff$has_changes()}}
\item \href{#method-GraphDiff-summary}{\code{GraphDiff$summary()}}
\item \href{#method-GraphDiff-clone}{\code{GraphDiff$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-GraphDiff-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphDiff$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphDiff-has_changes"></a>}}
\if{latex}{\out{\hypertarget{method-GraphDiff-has_changes}{}}}
\subsection{Method \code{has_changes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphDiff$has_changes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphDiff-summary"></a>}}
\if{latex}{\out{\hypertarget{method-GraphDiff-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphDiff$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GraphDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
