% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime-versioning.R
\name{GraphVersionManager}
\alias{GraphVersionManager}
\title{Graph Version Manager}
\description{
Manages versions of the reactive graph
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GraphVersionManager-new}{\code{GraphVersionManager$new()}}
\item \href{#method-GraphVersionManager-save_version}{\code{GraphVersionManager$save_version()}}
\item \href{#method-GraphVersionManager-get_version}{\code{GraphVersionManager$get_version()}}
\item \href{#method-GraphVersionManager-get_current}{\code{GraphVersionManager$get_current()}}
\item \href{#method-GraphVersionManager-get_previous}{\code{GraphVersionManager$get_previous()}}
\item \href{#method-GraphVersionManager-compare_versions}{\code{GraphVersionManager$compare_versions()}}
\item \href{#method-GraphVersionManager-clone}{\code{GraphVersionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-save_version"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-save_version}{}}}
\subsection{Method \code{save_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$save_version(graph, metadata = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-get_version}{}}}
\subsection{Method \code{get_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$get_version(version)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-get_current"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-get_current}{}}}
\subsection{Method \code{get_current()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$get_current()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-get_previous"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-get_previous}{}}}
\subsection{Method \code{get_previous()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$get_previous()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-compare_versions"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-compare_versions}{}}}
\subsection{Method \code{compare_versions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$compare_versions(v1, v2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVersionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVersionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVersionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
