% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-time-travel.R
\name{GraphVisualizer}
\alias{GraphVisualizer}
\title{Graph Visualizer}
\description{
Visualizes the reactive graph structure
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GraphVisualizer-new}{\code{GraphVisualizer$new()}}
\item \href{#method-GraphVisualizer-to_dot}{\code{GraphVisualizer$to_dot()}}
\item \href{#method-GraphVisualizer-to_json}{\code{GraphVisualizer$to_json()}}
\item \href{#method-GraphVisualizer-clone}{\code{GraphVisualizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVisualizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVisualizer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVisualizer$new(graph)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVisualizer-to_dot"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVisualizer-to_dot}{}}}
\subsection{Method \code{to_dot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVisualizer$to_dot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVisualizer-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVisualizer-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVisualizer$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GraphVisualizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GraphVisualizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphVisualizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
