% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-dynamic.R
\name{hideTab}
\alias{hideTab}
\title{Hide a tab}
\usage{
hideTab(inputId, target, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Tabset panel ID}

\item{target}{ID or value of tab to hide}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Hide a tab
}
