% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-reload-engine.R
\name{HotReloadEngine}
\alias{HotReloadEngine}
\title{Hot Reload Engine}
\description{
Manages hot reload process
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HotReloadEngine-new}{\code{HotReloadEngine$new()}}
\item \href{#method-HotReloadEngine-enable}{\code{HotReloadEngine$enable()}}
\item \href{#method-HotReloadEngine-disable}{\code{HotReloadEngine$disable()}}
\item \href{#method-HotReloadEngine-handle_file_change}{\code{HotReloadEngine$handle_file_change()}}
\item \href{#method-HotReloadEngine-reload}{\code{HotReloadEngine$reload()}}
\item \href{#method-HotReloadEngine-get_status}{\code{HotReloadEngine$get_status()}}
\item \href{#method-HotReloadEngine-clone}{\code{HotReloadEngine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$new(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-enable"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-enable}{}}}
\subsection{Method \code{enable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$enable(watch_paths = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-disable"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-disable}{}}}
\subsection{Method \code{disable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$disable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-handle_file_change"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-handle_file_change}{}}}
\subsection{Method \code{handle_file_change()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$handle_file_change(file_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-reload"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-get_status}{}}}
\subsection{Method \code{get_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$get_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotReloadEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HotReloadEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotReloadEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
