% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{HotShinyApp}
\alias{HotShinyApp}
\title{HotShiny App}
\description{
Main application object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HotShinyApp-new}{\code{HotShinyApp$new()}}
\item \href{#method-HotShinyApp-runApp}{\code{HotShinyApp$runApp()}}
\item \href{#method-HotShinyApp-start_http_server}{\code{HotShinyApp$start_http_server()}}
\item \href{#method-HotShinyApp-render_ui}{\code{HotShinyApp$render_ui()}}
\item \href{#method-HotShinyApp-render_app_html}{\code{HotShinyApp$render_app_html()}}
\item \href{#method-HotShinyApp-ui_to_html}{\code{HotShinyApp$ui_to_html()}}
\item \href{#method-HotShinyApp-get_graph}{\code{HotShinyApp$get_graph()}}
\item \href{#method-HotShinyApp-get_executor}{\code{HotShinyApp$get_executor()}}
\item \href{#method-HotShinyApp-get_state_manager}{\code{HotShinyApp$get_state_manager()}}
\item \href{#method-HotShinyApp-clone}{\code{HotShinyApp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-new"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$new(ui, server, builder, executor, state_manager, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-runApp"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-runApp}{}}}
\subsection{Method \code{runApp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$runApp(host = "127.0.0.1", port = 3838, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-start_http_server"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-start_http_server}{}}}
\subsection{Method \code{start_http_server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$start_http_server(host, port, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-render_ui"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-render_ui}{}}}
\subsection{Method \code{render_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$render_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-render_app_html"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-render_app_html}{}}}
\subsection{Method \code{render_app_html()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$render_app_html()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-ui_to_html"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-ui_to_html}{}}}
\subsection{Method \code{ui_to_html()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$ui_to_html(ui_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$get_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-get_executor"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-get_executor}{}}}
\subsection{Method \code{get_executor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$get_executor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-get_state_manager"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-get_state_manager}{}}}
\subsection{Method \code{get_state_manager()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$get_state_manager()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HotShinyApp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HotShinyApp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotShinyApp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
