% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{HTML}
\alias{HTML}
\title{Mark a character string as HTML}
\usage{
HTML(text)
}
\arguments{
\item{text}{Character string containing HTML}
}
\value{
An object of class "html" and "character"
}
\description{
This function marks a character string as HTML so it won't be escaped
when rendered.
}
