% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{htmlDependency}
\alias{htmlDependency}
\title{Create an HTML dependency}
\usage{
htmlDependency(
  name,
  version,
  src,
  script = NULL,
  stylesheet = NULL,
  head = NULL,
  attachment = NULL,
  package = NULL,
  all_files = TRUE
)
}
\arguments{
\item{name}{Dependency name}

\item{version}{Version string}

\item{src}{Source directory}

\item{script}{JavaScript files}

\item{stylesheet}{CSS files}

\item{head}{Additional head content}

\item{attachment}{Attached files}

\item{package}{Package name}

\item{all_files}{Include all files}
}
\value{
htmlDependency object
}
\description{
Create an HTML dependency
}
