% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{htmlEscape}
\alias{htmlEscape}
\title{Escape HTML special characters}
\usage{
htmlEscape(text, attribute = FALSE)
}
\arguments{
\item{text}{Text to escape}

\item{attribute}{Whether escaping for attribute value}
}
\value{
Escaped text
}
\description{
Escape HTML special characters
}
