% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{htmlOutput}
\alias{htmlOutput}
\title{Create an HTML output element}
\usage{
htmlOutput(
  outputId,
  inline = FALSE,
  container = if (inline) span else div,
  fill = FALSE,
  ...
)
}
\arguments{
\item{outputId}{Output variable to read the value from}

\item{inline}{Use inline element?}

\item{container}{Container function}

\item{fill}{Whether to allow filling}

\item{...}{Additional attributes}
}
\value{
An HTML output tag
}
\description{
Renders arbitrary HTML from the server.
}
