% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{htmlTemplate}
\alias{htmlTemplate}
\title{Process an HTML template}
\usage{
htmlTemplate(filename, ..., document_ = "auto")
}
\arguments{
\item{filename}{Path to template file}

\item{...}{Named values to substitute in template}

\item{document_}{Wrap in full HTML document}
}
\value{
Processed HTML
}
\description{
Process an HTML template
}
