% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{icon}
\alias{icon}
\title{Create an icon}
\usage{
icon(name, class = NULL, lib = "font-awesome", ...)
}
\arguments{
\item{name}{Icon name (e.g., "search", "check")}

\item{class}{Additional CSS classes}

\item{lib}{Icon library: "font-awesome", "bootstrap-icons", "glyphicon"}

\item{...}{Additional attributes}
}
\value{
An icon tag
}
\description{
Create an icon
}
