% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{imageOutput}
\alias{imageOutput}
\title{Create an image output element}
\usage{
imageOutput(
  outputId,
  width = "100\%",
  height = "400px",
  click = NULL,
  dblclick = NULL,
  hover = NULL,
  brush = NULL,
  inline = FALSE
)
}
\arguments{
\item{outputId}{Output variable for the image}

\item{width}{Width of the image}

\item{height}{Height of the image}

\item{click}{Click event options}

\item{dblclick}{Double-click event options}

\item{hover}{Hover event options}

\item{brush}{Brush event options}

\item{inline}{Display inline?}
}
\value{
An image output tag
}
\description{
Create an image output element
}
