% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{includeCSS}
\alias{includeCSS}
\title{Include CSS from a file}
\usage{
includeCSS(path, ...)
}
\arguments{
\item{path}{Path to CSS file}

\item{...}{Additional attributes for style tag}
}
\value{
Style tag with CSS content
}
\description{
Include CSS from a file
}
