% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{includeMarkdown}
\alias{includeMarkdown}
\title{Include Markdown from a file}
\usage{
includeMarkdown(path)
}
\arguments{
\item{path}{Path to markdown file}
}
\value{
Rendered HTML
}
\description{
Renders markdown to HTML.
}
