% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{includeScript}
\alias{includeScript}
\title{Include JavaScript from a file}
\usage{
includeScript(path, ...)
}
\arguments{
\item{path}{Path to JavaScript file}

\item{...}{Additional attributes for script tag}
}
\value{
Script tag with JS content
}
\description{
Include JavaScript from a file
}
