% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{includeText}
\alias{includeText}
\title{Include text from a file}
\usage{
includeText(path)
}
\arguments{
\item{path}{Path to text file}
}
\value{
Text content (escaped)
}
\description{
Include text from a file
}
