% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{incProgress}
\alias{incProgress}
\title{Increment progress value}
\usage{
incProgress(
  amount = 0.1,
  message = NULL,
  detail = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{amount}{Amount to increment (0-1 scale)}

\item{message}{Optional message}

\item{detail}{Optional detail}

\item{session}{Shiny session}
}
\value{
NULL (called for side effect)
}
\description{
Increment progress value
}
