% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-node-types.R
\name{InputNode}
\alias{InputNode}
\title{Input node - represents input$x}
\description{
Input node - represents input$x

Input node - represents input$x
}
\section{Super class}{
\code{\link[hotShiny:ReactiveNode]{hotShiny::ReactiveNode}} -> \code{InputNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InputNode-new}{\code{InputNode$new()}}
\item \href{#method-InputNode-clone}{\code{InputNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="compute_hash"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-compute_hash'><code>hotShiny::ReactiveNode$compute_hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="to_list"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-to_list'><code>hotShiny::ReactiveNode$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-InputNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputNode$new(id, input_name, version = 1L, source = NULL, env = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InputNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InputNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InputNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
