% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{inputPanel}
\alias{inputPanel}
\title{Create an input panel}
\usage{
inputPanel(...)
}
\arguments{
\item{...}{UI elements (inputs)}
}
\value{
An input panel tag
}
\description{
A panel for grouping inputs.
}
