% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-dynamic.R
\name{insertTab}
\alias{insertTab}
\title{Insert a tab into a tabset panel}
\usage{
insertTab(
  inputId,
  tab,
  target = NULL,
  position = c("after", "before"),
  select = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{Tabset panel ID}

\item{tab}{Tab panel to insert (from tabPanel())}

\item{target}{ID of tab to insert relative to}

\item{position}{Position relative to target: "before" or "after"}

\item{select}{Select the new tab?}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Insert a tab into a tabset panel
}
