% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{mainPanel}
\alias{mainPanel}
\title{Create a main panel}
\usage{
mainPanel(..., width = 8)
}
\arguments{
\item{...}{UI elements}

\item{width}{Width of main panel (1-12)}
}
\value{
A main panel tag
}
\description{
Create a main panel
}
