% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{makeTag}
\alias{makeTag}
\title{Create a tag function for a specific HTML element}
\usage{
makeTag(name)
}
\arguments{
\item{name}{HTML tag name}
}
\value{
Function that creates tags of that type
}
\description{
Create a tag function for a specific HTML element
}
