% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{markdown}
\alias{markdown}
\title{Render inline Markdown}
\usage{
markdown(text, ...)
}
\arguments{
\item{text}{Markdown text}

\item{...}{Additional arguments passed to markdown renderer}
}
\value{
Rendered HTML
}
\description{
Render inline Markdown
}
