% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-message-handler.R
\name{MessageHandler}
\alias{MessageHandler}
\title{Message Handler}
\description{
Processes and routes WebSocket messages
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessageHandler-new}{\code{MessageHandler$new()}}
\item \href{#method-MessageHandler-register}{\code{MessageHandler$register()}}
\item \href{#method-MessageHandler-handle}{\code{MessageHandler$handle()}}
\item \href{#method-MessageHandler-clone}{\code{MessageHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-MessageHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageHandler$new(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageHandler-register"></a>}}
\if{latex}{\out{\hypertarget{method-MessageHandler-register}{}}}
\subsection{Method \code{register()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageHandler$register(message_type, handler)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageHandler-handle"></a>}}
\if{latex}{\out{\hypertarget{method-MessageHandler-handle}{}}}
\subsection{Method \code{handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageHandler$handle(message, ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessageHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
