% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-state-preservation.R
\name{migrate_modified_state}
\alias{migrate_modified_state}
\title{Migrate state for modified nodes}
\usage{
migrate_modified_state(
  diff,
  old_graph,
  new_graph,
  state_manager,
  preservation_manager
)
}
\arguments{
\item{diff}{GraphDiff result}

\item{old_graph}{Old graph}

\item{new_graph}{New graph}

\item{state_manager}{StateManager}

\item{preservation_manager}{StatePreservationManager}
}
\description{
Attempts to migrate state when nodes are modified
}
