% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{modalButton}
\alias{modalButton}
\title{Create a modal button}
\usage{
modalButton(label, icon = NULL)
}
\arguments{
\item{label}{Button label}

\item{icon}{Optional icon}
}
\value{
A modal button tag
}
\description{
Creates a button that dismisses the modal.
}
