% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{modalDialog}
\alias{modalDialog}
\title{Create a modal dialog UI}
\usage{
modalDialog(
  ...,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l", "xl"),
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{...}{UI elements for modal body}

\item{title}{Modal title}

\item{footer}{Modal footer content (default: OK and Cancel buttons)}

\item{size}{Modal size: "s", "m", "l", "xl"}

\item{easyClose}{Close on click outside or Escape?}

\item{fade}{Use fade animation?}
}
\value{
A modal dialog tag
}
\description{
Create a modal dialog UI
}
