% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-navigation.R
\name{navbarMenu}
\alias{navbarMenu}
\title{Create a dropdown menu for navbar}
\usage{
navbarMenu(title, ..., menuName = title, icon = NULL)
}
\arguments{
\item{title}{Menu title}

\item{...}{Tab panels or menu items}

\item{menuName}{Internal name for menu}

\item{icon}{Optional icon}
}
\value{
A navbar menu tag
}
\description{
Create a dropdown menu for navbar
}
