% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-navigation.R
\name{navlistPanel}
\alias{navlistPanel}
\title{Create a navigation list panel}
\usage{
navlistPanel(
  ...,
  id = NULL,
  selected = NULL,
  well = TRUE,
  fluid = TRUE,
  widths = c(4, 8)
)
}
\arguments{
\item{...}{Tab panels}

\item{id}{Optional id}

\item{selected}{Initially selected tab}

\item{well}{Display in well panel?}

\item{fluid}{Use fluid layout?}

\item{widths}{Widths for nav and content columns}
}
\value{
A navlist panel tag
}
\description{
Vertical navigation with content panels.
}
