% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-node-types.R
\name{new_node_id}
\alias{new_node_id}
\title{Create a new node ID}
\usage{
new_node_id(prefix = "node", counter = NULL)
}
\arguments{
\item{prefix}{Prefix for the ID}

\item{counter}{Counter for uniqueness}
}
\description{
Create a new node ID
}
