% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-graph-diff.R
\name{node_unchanged}
\alias{node_unchanged}
\title{Check if a node is unchanged between versions}
\usage{
node_unchanged(old_node, new_node)
}
\arguments{
\item{old_node}{Old node}

\item{new_node}{New node}
}
\value{
TRUE if unchanged
}
\description{
Check if a node is unchanged between versions
}
