% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{normalizeChoices}
\alias{normalizeChoices}
\title{Normalize choices to named list}
\usage{
normalizeChoices(choices)
}
\arguments{
\item{choices}{Vector, list, or named vector of choices}
}
\value{
Named list
}
\description{
Normalize choices to named list
}
