% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{numericInput}
\alias{numericInput}
\title{Create a numeric input control}
\usage{
numericInput(
  inputId,
  label,
  value,
  min = NA,
  max = NA,
  step = NA,
  width = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}

\item{width}{Width of the input}
}
\value{
A numeric input tag
}
\description{
Create a numeric input control
}
