% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-node-types.R
\name{ObserverNode}
\alias{ObserverNode}
\title{Observer node - represents observe({ ... })}
\description{
Observer node - represents observe({ ... })

Observer node - represents observe({ ... })
}
\section{Super class}{
\code{\link[hotShiny:ReactiveNode]{hotShiny::ReactiveNode}} -> \code{ObserverNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObserverNode-new}{\code{ObserverNode$new()}}
\item \href{#method-ObserverNode-clone}{\code{ObserverNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="compute_hash"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-compute_hash'><code>hotShiny::ReactiveNode$compute_hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="to_list"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-to_list'><code>hotShiny::ReactiveNode$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverNode$new(
  id,
  deps = list(),
  expr = NULL,
  priority = 0L,
  once = FALSE,
  suspended = FALSE,
  version = 1L,
  source = NULL,
  env = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
