% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-observe.R
\name{ObserverProxy}
\alias{ObserverProxy}
\title{Observer Proxy}
\description{
Proxy object for observers
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObserverProxy-new}{\code{ObserverProxy$new()}}
\item \href{#method-ObserverProxy-suspend}{\code{ObserverProxy$suspend()}}
\item \href{#method-ObserverProxy-resume}{\code{ObserverProxy$resume()}}
\item \href{#method-ObserverProxy-destroy}{\code{ObserverProxy$destroy()}}
\item \href{#method-ObserverProxy-clone}{\code{ObserverProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverProxy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverProxy$new(node_id, builder)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverProxy-suspend"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverProxy-suspend}{}}}
\subsection{Method \code{suspend()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverProxy$suspend()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverProxy-resume"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverProxy-resume}{}}}
\subsection{Method \code{resume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverProxy$resume()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverProxy-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverProxy-destroy}{}}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverProxy$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObserverProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObserverProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObserverProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
