% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-node-types.R
\name{OutputNode}
\alias{OutputNode}
\title{Output node - represents output$x}
\description{
Output node - represents output$x

Output node - represents output$x
}
\section{Super class}{
\code{\link[hotShiny:ReactiveNode]{hotShiny::ReactiveNode}} -> \code{OutputNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OutputNode-new}{\code{OutputNode$new()}}
\item \href{#method-OutputNode-clone}{\code{OutputNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="compute_hash"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-compute_hash'><code>hotShiny::ReactiveNode$compute_hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="to_list"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-to_list'><code>hotShiny::ReactiveNode$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-OutputNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputNode$new(
  id,
  output_name,
  render_type = NULL,
  deps = list(),
  expr = NULL,
  version = 1L,
  source = NULL,
  env = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OutputNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
