% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{outputOptions}
\alias{outputOptions}
\title{Set options for an output object}
\usage{
outputOptions(x, name, ...)
}
\arguments{
\item{x}{Render function output}

\item{name}{Name of the option}

\item{...}{Option values}
}
\value{
Modified output with options
}
\description{
Set options for an output object
}
