% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{passwordInput}
\alias{passwordInput}
\title{Create a password input control}
\usage{
passwordInput(inputId, label, value = "", width = NULL, placeholder = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{width}{Width of the input}

\item{placeholder}{Placeholder text}
}
\value{
A password input tag
}
\description{
Create a password input control
}
