% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{plotOutput}
\alias{plotOutput}
\title{Create a plot output element}
\usage{
plotOutput(
  outputId,
  width = "100\%",
  height = "400px",
  click = NULL,
  dblclick = NULL,
  hover = NULL,
  brush = NULL,
  inline = FALSE
)
}
\arguments{
\item{outputId}{Output variable for the plot}

\item{width}{Width of the plot}

\item{height}{Height of the plot}

\item{click}{Click event options (NULL or clickOpts)}

\item{dblclick}{Double-click event options}

\item{hover}{Hover event options}

\item{brush}{Brush event options}

\item{inline}{Display inline?}
}
\value{
A plot output tag
}
\description{
Create a plot output element
}
