% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-dynamic.R
\name{prependTab}
\alias{prependTab}
\title{Prepend a tab to a tabset panel}
\usage{
prependTab(inputId, tab, select = FALSE, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Tabset panel ID}

\item{tab}{Tab panel to prepend}

\item{select}{Select the new tab?}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Prepend a tab to a tabset panel
}
