% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-reactive.R
\name{reactive}
\alias{reactive}
\title{Create a reactive expression}
\usage{
reactive(x, label = NULL)
}
\arguments{
\item{x}{Expression to make reactive}

\item{label}{Optional label for debugging}
}
\value{
Reactive proxy object
}
\description{
Create a reactive expression
}
