% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime-executor.R
\name{ReactiveExecutor}
\alias{ReactiveExecutor}
\title{Reactive Executor}
\description{
Executes reactive graph nodes in correct order
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveExecutor-set_app}{\code{ReactiveExecutor$set_app()}}
\item \href{#method-ReactiveExecutor-new}{\code{ReactiveExecutor$new()}}
\item \href{#method-ReactiveExecutor-load_helper_functions}{\code{ReactiveExecutor$load_helper_functions()}}
\item \href{#method-ReactiveExecutor-get_helper_function}{\code{ReactiveExecutor$get_helper_function()}}
\item \href{#method-ReactiveExecutor-execute}{\code{ReactiveExecutor$execute()}}
\item \href{#method-ReactiveExecutor-should_execute}{\code{ReactiveExecutor$should_execute()}}
\item \href{#method-ReactiveExecutor-execute_node}{\code{ReactiveExecutor$execute_node()}}
\item \href{#method-ReactiveExecutor-create_input_proxy}{\code{ReactiveExecutor$create_input_proxy()}}
\item \href{#method-ReactiveExecutor-execute_reactive}{\code{ReactiveExecutor$execute_reactive()}}
\item \href{#method-ReactiveExecutor-execute_observer}{\code{ReactiveExecutor$execute_observer()}}
\item \href{#method-ReactiveExecutor-execute_render}{\code{ReactiveExecutor$execute_render()}}
\item \href{#method-ReactiveExecutor-get_value}{\code{ReactiveExecutor$get_value()}}
\item \href{#method-ReactiveExecutor-set_input}{\code{ReactiveExecutor$set_input()}}
\item \href{#method-ReactiveExecutor-send_output_updates}{\code{ReactiveExecutor$send_output_updates()}}
\item \href{#method-ReactiveExecutor-get_app}{\code{ReactiveExecutor$get_app()}}
\item \href{#method-ReactiveExecutor-get_state_manager}{\code{ReactiveExecutor$get_state_manager()}}
\item \href{#method-ReactiveExecutor-clone}{\code{ReactiveExecutor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-set_app"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-set_app}{}}}
\subsection{Method \code{set_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$set_app(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$new(graph, state_manager = NULL, builder = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-load_helper_functions"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-load_helper_functions}{}}}
\subsection{Method \code{load_helper_functions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$load_helper_functions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-get_helper_function"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-get_helper_function}{}}}
\subsection{Method \code{get_helper_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$get_helper_function(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-execute"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-execute}{}}}
\subsection{Method \code{execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$execute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-should_execute"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-should_execute}{}}}
\subsection{Method \code{should_execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$should_execute(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-execute_node"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-execute_node}{}}}
\subsection{Method \code{execute_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$execute_node(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-create_input_proxy"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-create_input_proxy}{}}}
\subsection{Method \code{create_input_proxy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$create_input_proxy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-execute_reactive"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-execute_reactive}{}}}
\subsection{Method \code{execute_reactive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$execute_reactive(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-execute_observer"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-execute_observer}{}}}
\subsection{Method \code{execute_observer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$execute_observer(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-execute_render"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-execute_render}{}}}
\subsection{Method \code{execute_render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$execute_render(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$get_value(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-set_input"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-set_input}{}}}
\subsection{Method \code{set_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$set_input(input_name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-send_output_updates"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-send_output_updates}{}}}
\subsection{Method \code{send_output_updates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$send_output_updates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-get_app"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-get_app}{}}}
\subsection{Method \code{get_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$get_app()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-get_state_manager"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-get_state_manager}{}}}
\subsection{Method \code{get_state_manager()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$get_state_manager()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveExecutor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveExecutor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveExecutor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
