% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-graph-builder.R
\name{ReactiveGraph}
\alias{ReactiveGraph}
\title{Reactive Graph}
\description{
Container for the reactive graph structure
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveGraph-new}{\code{ReactiveGraph$new()}}
\item \href{#method-ReactiveGraph-add_node}{\code{ReactiveGraph$add_node()}}
\item \href{#method-ReactiveGraph-get_node}{\code{ReactiveGraph$get_node()}}
\item \href{#method-ReactiveGraph-get_all_nodes}{\code{ReactiveGraph$get_all_nodes()}}
\item \href{#method-ReactiveGraph-get_nodes_by_type}{\code{ReactiveGraph$get_nodes_by_type()}}
\item \href{#method-ReactiveGraph-to_list}{\code{ReactiveGraph$to_list()}}
\item \href{#method-ReactiveGraph-topological_sort}{\code{ReactiveGraph$topological_sort()}}
\item \href{#method-ReactiveGraph-clone}{\code{ReactiveGraph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$new(version = 1L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-add_node"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-add_node}{}}}
\subsection{Method \code{add_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$add_node(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-get_node"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-get_node}{}}}
\subsection{Method \code{get_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$get_node(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-get_all_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-get_all_nodes}{}}}
\subsection{Method \code{get_all_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$get_all_nodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-get_nodes_by_type"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-get_nodes_by_type}{}}}
\subsection{Method \code{get_nodes_by_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$get_nodes_by_type(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-topological_sort"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-topological_sort}{}}}
\subsection{Method \code{topological_sort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$topological_sort()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveGraph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveGraph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveGraph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
