% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-node-types.R
\name{ReactiveNode}
\alias{ReactiveNode}
\title{Base class for all reactive nodes}
\description{
Base class for all reactive nodes

Base class for all reactive nodes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveNode-new}{\code{ReactiveNode$new()}}
\item \href{#method-ReactiveNode-to_list}{\code{ReactiveNode$to_list()}}
\item \href{#method-ReactiveNode-compute_hash}{\code{ReactiveNode$compute_hash()}}
\item \href{#method-ReactiveNode-clone}{\code{ReactiveNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveNode$new(
  id,
  type,
  deps = list(),
  expr = NULL,
  version = 1L,
  source = NULL,
  metadata = list(),
  env = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier for the node}

\item{\code{type}}{Node type (reactive, input, output, observe, render)}

\item{\code{deps}}{List of dependency node IDs}

\item{\code{expr}}{Expression AST (not closure)}

\item{\code{version}}{Version number for hot reload}

\item{\code{source}}{Source location (file, line)}

\item{\code{metadata}}{Additional metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveNode-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveNode-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveNode$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveNode-compute_hash"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveNode-compute_hash}{}}}
\subsection{Method \code{compute_hash()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveNode$compute_hash()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
