% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-reactive.R
\name{ReactiveProxy}
\alias{ReactiveProxy}
\title{Reactive Proxy}
\description{
Proxy object that mimics Shiny's reactive behavior
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveProxy-new}{\code{ReactiveProxy$new()}}
\item \href{#method-ReactiveProxy-get_executor}{\code{ReactiveProxy$get_executor()}}
\item \href{#method-ReactiveProxy-call}{\code{ReactiveProxy$call()}}
\item \href{#method-ReactiveProxy-clone}{\code{ReactiveProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveProxy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveProxy$new(node_id, builder)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveProxy-get_executor"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveProxy-get_executor}{}}}
\subsection{Method \code{get_executor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveProxy$get_executor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveProxy-call"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveProxy-call}{}}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveProxy$call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
