% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-values.R
\name{reactiveValues}
\alias{reactiveValues}
\title{Create reactive values}
\usage{
reactiveValues(...)
}
\arguments{
\item{...}{Named values}
}
\value{
ReactiveValues object
}
\description{
Create reactive values
}
