% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-serializer.R
\name{reconstruct_node}
\alias{reconstruct_node}
\title{Reconstruct a node from serialized data}
\usage{
reconstruct_node(node_data)
}
\arguments{
\item{node_data}{List representation of node}
}
\value{
ReactiveNode object
}
\description{
Reconstruct a node from serialized data
}
