% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{removeModal}
\alias{removeModal}
\title{Remove a modal dialog}
\usage{
removeModal(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Remove a modal dialog
}
