% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{removeNotification}
\alias{removeNotification}
\title{Remove a notification}
\usage{
removeNotification(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Notification ID}

\item{session}{Shiny session}
}
\value{
NULL (called for side effect)
}
\description{
Remove a notification
}
