% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-dynamic.R
\name{removeTab}
\alias{removeTab}
\title{Remove a tab from a tabset panel}
\usage{
removeTab(inputId, target, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Tabset panel ID}

\item{target}{ID or value of tab to remove}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Remove a tab from a tabset panel
}
