% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{render_impl}
\alias{render_impl}
\title{Generic render implementation}
\usage{
render_impl(
  render_type,
  expr,
  env,
  quoted,
  outputArgs,
  output_name = NULL,
  ...
)
}
\arguments{
\item{render_type}{Type of render function}

\item{expr}{Expression to render}

\item{env}{Environment}

\item{quoted}{Whether expression is quoted}

\item{outputArgs}{Output arguments}

\item{...}{Additional arguments}
}
\description{
Generic render implementation
}
