% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{renderCachedPlot}
\alias{renderCachedPlot}
\title{Render cached plot output}
\usage{
renderCachedPlot(
  expr,
  cacheKeyExpr,
  sizePolicy = sizeGrowthRatio(width = 400, height = 400, growthRate = 1.2),
  res = 72,
  cache = "app",
  ...
)
}
\arguments{
\item{expr}{Expression to render the plot}

\item{cacheKeyExpr}{Expression that returns the cache key}

\item{sizePolicy}{Size policy function}

\item{res}{Resolution}

\item{cache}{Cache object}

\item{...}{Additional arguments passed to renderPlot}
}
\description{
Like renderPlot but with caching support.
}
