% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{renderDataTable}
\alias{renderDataTable}
\title{Render data table output}
\usage{
renderDataTable(expr, env = parent.frame(), quoted = FALSE, ...)
}
\arguments{
\item{expr}{Expression to render}

\item{env}{Environment}

\item{quoted}{Whether expression is quoted}

\item{...}{Additional arguments}
}
\description{
Render data table output
}
