% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{renderImage}
\alias{renderImage}
\title{Render image output}
\usage{
renderImage(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  deleteFile = TRUE,
  outputArgs = list(),
  ...
)
}
\arguments{
\item{expr}{Expression that returns a list with 'src' (file path) and optionally
'contentType', 'width', 'height', 'alt'}

\item{env}{Environment for evaluation}

\item{quoted}{Whether expression is quoted}

\item{deleteFile}{Delete file after sending?}

\item{outputArgs}{Output arguments}

\item{...}{Additional arguments}
}
\description{
Renders an image from a file.
}
