% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-node-types.R
\name{RenderNode}
\alias{RenderNode}
\title{Render node - represents render*() functions}
\description{
Render node - represents render*() functions

Render node - represents render*() functions
}
\section{Super class}{
\code{\link[hotShiny:ReactiveNode]{hotShiny::ReactiveNode}} -> \code{RenderNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RenderNode-new}{\code{RenderNode$new()}}
\item \href{#method-RenderNode-clone}{\code{RenderNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="compute_hash"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-compute_hash'><code>hotShiny::ReactiveNode$compute_hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hotShiny" data-topic="ReactiveNode" data-id="to_list"><a href='../../hotShiny/html/ReactiveNode.html#method-ReactiveNode-to_list'><code>hotShiny::ReactiveNode$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-RenderNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderNode$new(
  id,
  render_type,
  output_name,
  deps = list(),
  expr = NULL,
  version = 1L,
  source = NULL,
  env = NULL,
  outputArgs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RenderNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
