% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{renderPlot}
\alias{renderPlot}
\title{Render plot output}
\usage{
renderPlot(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  width = "auto",
  height = "auto",
  ...
)
}
\arguments{
\item{expr}{Expression to render}

\item{env}{Environment}

\item{quoted}{Whether expression is quoted}

\item{width}{Width}

\item{height}{Height}

\item{...}{Additional arguments}
}
\description{
Render plot output
}
