% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{renderPrint}
\alias{renderPrint}
\title{Render print output}
\usage{
renderPrint(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  width = getOption("width"),
  outputArgs = list(),
  ...
)
}
\arguments{
\item{expr}{Expression to evaluate and capture print output}

\item{env}{Environment for evaluation}

\item{quoted}{Whether expression is quoted}

\item{width}{Print width}

\item{outputArgs}{Output arguments}

\item{...}{Additional arguments}
}
\description{
Captures printed output from an expression.
}
