% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{RenderProxy}
\alias{RenderProxy}
\title{Render Proxy}
\description{
Proxy object for render functions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RenderProxy-new}{\code{RenderProxy$new()}}
\item \href{#method-RenderProxy-set_output_name}{\code{RenderProxy$set_output_name()}}
\item \href{#method-RenderProxy-clone}{\code{RenderProxy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderProxy-new"></a>}}
\if{latex}{\out{\hypertarget{method-RenderProxy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderProxy$new(node_id, render_type, builder)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderProxy-set_output_name"></a>}}
\if{latex}{\out{\hypertarget{method-RenderProxy-set_output_name}{}}}
\subsection{Method \code{set_output_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderProxy$set_output_name(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderProxy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RenderProxy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
