% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{renderText}
\alias{renderText}
\title{Render text output}
\usage{
renderText(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  outputArgs = list(),
  output_name = NULL,
  ...
)
}
\arguments{
\item{expr}{Expression to render}

\item{env}{Environment}

\item{quoted}{Whether expression is quoted}

\item{outputArgs}{Output arguments}

\item{output_name}{Optional output name (usually set automatically)}

\item{...}{Additional arguments}
}
\description{
Render text output
}
