% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{selectInput}
\alias{selectInput}
\title{Create a select list input control}
\usage{
selectInput(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = NULL,
  size = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{choices}{List of values to select from}

\item{selected}{Initially selected value(s)}

\item{multiple}{Allow multiple selections?}

\item{selectize}{Use selectize.js?}

\item{width}{Width of the input}

\item{size}{Number of visible options (for non-selectize)}
}
\value{
A select input tag
}
\description{
Create a select list input control
}
