% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{selectizeInput}
\alias{selectizeInput}
\title{Create a selectize input (enhanced select)}
\usage{
selectizeInput(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  options = NULL,
  width = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{choices}{List of values to select from}

\item{selected}{Initially selected value(s)}

\item{multiple}{Allow multiple selections?}

\item{options}{Options for selectize.js}

\item{width}{Width of the input}
}
\value{
A selectize input tag
}
\description{
Create a selectize input (enhanced select)
}
