% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir-serializer.R
\name{serialize_graph}
\alias{serialize_graph}
\title{Serialize graph to JSON}
\usage{
serialize_graph(graph, pretty = FALSE)
}
\arguments{
\item{graph}{ReactiveGraph object}

\item{pretty}{Whether to pretty-print JSON}
}
\value{
JSON string
}
\description{
Serialize graph to JSON
}
