% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-session-manager.R
\name{Session}
\alias{Session}
\title{Session}
\description{
Represents a client session
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Session-new}{\code{Session$new()}}
\item \href{#method-Session-update_activity}{\code{Session$update_activity()}}
\item \href{#method-Session-clone}{\code{Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-new"></a>}}
\if{latex}{\out{\hypertarget{method-Session-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$new(id, ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-update_activity"></a>}}
\if{latex}{\out{\hypertarget{method-Session-update_activity}{}}}
\subsection{Method \code{update_activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$update_activity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
