% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-session-manager.R
\name{SessionManager}
\alias{SessionManager}
\title{Session Manager}
\description{
Manages client sessions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionManager-new}{\code{SessionManager$new()}}
\item \href{#method-SessionManager-create_session}{\code{SessionManager$create_session()}}
\item \href{#method-SessionManager-get_session}{\code{SessionManager$get_session()}}
\item \href{#method-SessionManager-remove_session}{\code{SessionManager$remove_session()}}
\item \href{#method-SessionManager-get_all_sessions}{\code{SessionManager$get_all_sessions()}}
\item \href{#method-SessionManager-cleanup_inactive}{\code{SessionManager$cleanup_inactive()}}
\item \href{#method-SessionManager-clone}{\code{SessionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-create_session"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-create_session}{}}}
\subsection{Method \code{create_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$create_session(ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-get_session"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-get_session}{}}}
\subsection{Method \code{get_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$get_session(session_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-remove_session"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-remove_session}{}}}
\subsection{Method \code{remove_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$remove_session(session_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-get_all_sessions"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-get_all_sessions}{}}}
\subsection{Method \code{get_all_sessions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$get_all_sessions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-cleanup_inactive"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-cleanup_inactive}{}}}
\subsection{Method \code{cleanup_inactive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$cleanup_inactive(timeout_seconds = 3600)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
