% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{set_current_output_name}
\alias{set_current_output_name}
\title{Set current output name (called during output$x <- assignment)}
\usage{
set_current_output_name(name)
}
\arguments{
\item{name}{Output name}
}
\description{
Set current output name (called during output$x <- assignment)
}
