% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-reactive.R
\name{set_graph_builder}
\alias{set_graph_builder}
\title{Set the current graph builder}
\usage{
set_graph_builder(builder)
}
\arguments{
\item{builder}{GraphBuilder instance}
}
\description{
Set the current graph builder
}
