% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{setProgress}
\alias{setProgress}
\title{Set progress value}
\usage{
setProgress(
  value = NULL,
  message = NULL,
  detail = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{value}{Progress value}

\item{message}{Optional message}

\item{detail}{Optional detail}

\item{session}{Shiny session}
}
\value{
NULL (called for side effect)
}
\description{
Set progress value
}
