% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{showModal}
\alias{showModal}
\title{Show a modal dialog}
\usage{
showModal(ui, session = getDefaultReactiveDomain())
}
\arguments{
\item{ui}{Modal UI (from modalDialog)}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Show a modal dialog
}
