% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{showNotification}
\alias{showNotification}
\title{Show a notification}
\usage{
showNotification(
  ui,
  action = NULL,
  duration = 5,
  closeButton = TRUE,
  id = NULL,
  type = c("default", "message", "warning", "error"),
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{ui}{Notification content}

\item{action}{Optional action button/link}

\item{duration}{Duration in seconds (NULL for persistent)}

\item{closeButton}{Show close button?}

\item{id}{Unique ID (for removal)}

\item{type}{Type: "default", "message", "warning", "error"}

\item{session}{Shiny session}
}
\value{
The notification ID
}
\description{
Show a notification
}
