% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-dynamic.R
\name{showTab}
\alias{showTab}
\title{Show a hidden tab}
\usage{
showTab(inputId, target, select = FALSE, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Tabset panel ID}

\item{target}{ID or value of tab to show}

\item{select}{Select the tab after showing?}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Show a hidden tab
}
