% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{sidebarLayout}
\alias{sidebarLayout}
\title{Create a sidebar layout}
\usage{
sidebarLayout(
  sidebarPanel,
  mainPanel,
  position = c("left", "right"),
  fluid = TRUE
)
}
\arguments{
\item{sidebarPanel}{Sidebar panel content}

\item{mainPanel}{Main panel content}

\item{position}{Position of sidebar: "left" or "right"}

\item{fluid}{Use fluid widths?}
}
\value{
A sidebar layout tag
}
\description{
Create a sidebar layout
}
