% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{sidebarPanel}
\alias{sidebarPanel}
\title{Create a sidebar panel}
\usage{
sidebarPanel(..., width = 4)
}
\arguments{
\item{...}{UI elements}

\item{width}{Width of sidebar (1-12)}
}
\value{
A sidebar panel tag
}
\description{
Create a sidebar panel
}
