% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{singleton}
\alias{singleton}
\title{Include only once}
\usage{
singleton(x)
}
\arguments{
\item{x}{Content to mark as singleton}
}
\value{
Singleton-marked content
}
\description{
Marks content to be included only once in the HTML output,
even if the function is called multiple times.
}
