% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-render.R
\name{sizeGrowthRatio}
\alias{sizeGrowthRatio}
\title{Size growth ratio policy for cached plots}
\usage{
sizeGrowthRatio(width = 400, height = 400, growthRate = 1.2)
}
\arguments{
\item{width}{Base width}

\item{height}{Base height}

\item{growthRate}{Growth rate multiplier}
}
\value{
Size policy function
}
\description{
Size growth ratio policy for cached plots
}
