% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{sliderInput}
\alias{sliderInput}
\title{Create a slider input}
\usage{
sliderInput(
  inputId,
  label,
  min,
  max,
  value,
  step = NULL,
  round = FALSE,
  ticks = TRUE,
  animate = FALSE,
  width = NULL,
  sep = ",",
  pre = NULL,
  post = NULL,
  timeFormat = NULL,
  timezone = NULL,
  dragRange = TRUE
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{min}{Minimum value}

\item{max}{Maximum value}

\item{value}{Initial value (single or two values for range)}

\item{step}{Step size}

\item{round}{Round to nearest step?}

\item{ticks}{Show tick marks?}

\item{animate}{Add animation? Can be TRUE or animationOptions()}

\item{width}{Width of the slider}

\item{sep}{Thousands separator}

\item{pre}{Prefix for displayed values}

\item{post}{Suffix for displayed values}

\item{timeFormat}{Time format if values are dates}

\item{timezone}{Time zone}

\item{dragRange}{Allow dragging the range?}
}
\value{
A slider input tag
}
\description{
Create a slider input
}
