% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{splitLayout}
\alias{splitLayout}
\title{Create a split layout}
\usage{
splitLayout(..., cellWidths = NULL, cellArgs = list())
}
\arguments{
\item{...}{UI elements to split between}

\item{cellWidths}{Widths for each cell (vector or single value)}

\item{cellArgs}{Additional arguments for cells}
}
\value{
A split layout tag
}
\description{
Splits space evenly between elements.
}
