% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime-state-manager.R
\name{StateManager}
\alias{StateManager}
\title{State Manager}
\description{
Manages reactive values and application state
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StateManager-new}{\code{StateManager$new()}}
\item \href{#method-StateManager-set_value}{\code{StateManager$set_value()}}
\item \href{#method-StateManager-get_value}{\code{StateManager$get_value()}}
\item \href{#method-StateManager-mark_dirty}{\code{StateManager$mark_dirty()}}
\item \href{#method-StateManager-is_dirty}{\code{StateManager$is_dirty()}}
\item \href{#method-StateManager-clear_dirty}{\code{StateManager$clear_dirty()}}
\item \href{#method-StateManager-clear_all_dirty}{\code{StateManager$clear_all_dirty()}}
\item \href{#method-StateManager-get_dirty_nodes}{\code{StateManager$get_dirty_nodes()}}
\item \href{#method-StateManager-set_execution_state}{\code{StateManager$set_execution_state()}}
\item \href{#method-StateManager-get_execution_state}{\code{StateManager$get_execution_state()}}
\item \href{#method-StateManager-serialize_state}{\code{StateManager$serialize_state()}}
\item \href{#method-StateManager-restore_state}{\code{StateManager$restore_state()}}
\item \href{#method-StateManager-clear_all}{\code{StateManager$clear_all()}}
\item \href{#method-StateManager-clone}{\code{StateManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$set_value(node_id, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$get_value(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-mark_dirty"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-mark_dirty}{}}}
\subsection{Method \code{mark_dirty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$mark_dirty(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-is_dirty"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-is_dirty}{}}}
\subsection{Method \code{is_dirty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$is_dirty(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-clear_dirty"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-clear_dirty}{}}}
\subsection{Method \code{clear_dirty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$clear_dirty(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-clear_all_dirty"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-clear_all_dirty}{}}}
\subsection{Method \code{clear_all_dirty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$clear_all_dirty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-get_dirty_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-get_dirty_nodes}{}}}
\subsection{Method \code{get_dirty_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$get_dirty_nodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-set_execution_state"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-set_execution_state}{}}}
\subsection{Method \code{set_execution_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$set_execution_state(node_id, state)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-get_execution_state"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-get_execution_state}{}}}
\subsection{Method \code{get_execution_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$get_execution_state(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-serialize_state"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-serialize_state}{}}}
\subsection{Method \code{serialize_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$serialize_state(node_ids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-restore_state"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-restore_state}{}}}
\subsection{Method \code{restore_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$restore_state(state_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-clear_all"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-clear_all}{}}}
\subsection{Method \code{clear_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$clear_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StateManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
