% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-reload-state-preservation.R
\name{StatePreservationManager}
\alias{StatePreservationManager}
\title{State Preservation Manager}
\description{
Manages state preservation during hot reload
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatePreservationManager-new}{\code{StatePreservationManager$new()}}
\item \href{#method-StatePreservationManager-preserve_node_state}{\code{StatePreservationManager$preserve_node_state()}}
\item \href{#method-StatePreservationManager-preserve_nodes_state}{\code{StatePreservationManager$preserve_nodes_state()}}
\item \href{#method-StatePreservationManager-restore_node_state}{\code{StatePreservationManager$restore_node_state()}}
\item \href{#method-StatePreservationManager-restore_nodes_state}{\code{StatePreservationManager$restore_nodes_state()}}
\item \href{#method-StatePreservationManager-clear}{\code{StatePreservationManager$clear()}}
\item \href{#method-StatePreservationManager-clear_nodes}{\code{StatePreservationManager$clear_nodes()}}
\item \href{#method-StatePreservationManager-get_preserved_nodes}{\code{StatePreservationManager$get_preserved_nodes()}}
\item \href{#method-StatePreservationManager-clone}{\code{StatePreservationManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-preserve_node_state"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-preserve_node_state}{}}}
\subsection{Method \code{preserve_node_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$preserve_node_state(node_id, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-preserve_nodes_state"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-preserve_nodes_state}{}}}
\subsection{Method \code{preserve_nodes_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$preserve_nodes_state(node_ids, state_manager)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-restore_node_state"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-restore_node_state}{}}}
\subsection{Method \code{restore_node_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$restore_node_state(node_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-restore_nodes_state"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-restore_nodes_state}{}}}
\subsection{Method \code{restore_nodes_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$restore_nodes_state(node_ids, state_manager)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-clear"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-clear_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-clear_nodes}{}}}
\subsection{Method \code{clear_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$clear_nodes(node_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-get_preserved_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-get_preserved_nodes}{}}}
\subsection{Method \code{get_preserved_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$get_preserved_nodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatePreservationManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatePreservationManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatePreservationManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
