% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-strict-mode.R
\name{StrictModeChecker}
\alias{StrictModeChecker}
\title{Strict Mode Checker}
\description{
Checks for side effects and non-deterministic operations
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrictModeChecker-new}{\code{StrictModeChecker$new()}}
\item \href{#method-StrictModeChecker-check_expression}{\code{StrictModeChecker$check_expression()}}
\item \href{#method-StrictModeChecker-walk_ast}{\code{StrictModeChecker$walk_ast()}}
\item \href{#method-StrictModeChecker-get_warnings}{\code{StrictModeChecker$get_warnings()}}
\item \href{#method-StrictModeChecker-get_errors}{\code{StrictModeChecker$get_errors()}}
\item \href{#method-StrictModeChecker-clear}{\code{StrictModeChecker$clear()}}
\item \href{#method-StrictModeChecker-enable}{\code{StrictModeChecker$enable()}}
\item \href{#method-StrictModeChecker-disable}{\code{StrictModeChecker$disable()}}
\item \href{#method-StrictModeChecker-clone}{\code{StrictModeChecker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$new(enabled = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-check_expression"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-check_expression}{}}}
\subsection{Method \code{check_expression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$check_expression(expr, context = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-walk_ast"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-walk_ast}{}}}
\subsection{Method \code{walk_ast()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$walk_ast(expr, issues, context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-get_warnings"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-get_warnings}{}}}
\subsection{Method \code{get_warnings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$get_warnings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-get_errors"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-get_errors}{}}}
\subsection{Method \code{get_errors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$get_errors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-clear"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-enable"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-enable}{}}}
\subsection{Method \code{enable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$enable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-disable"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-disable}{}}}
\subsection{Method \code{disable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$disable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrictModeChecker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrictModeChecker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrictModeChecker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
