% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{submitButton}
\alias{submitButton}
\title{Create a submit button}
\usage{
submitButton(text = "Submit", icon = NULL, width = NULL)
}
\arguments{
\item{text}{Button text}

\item{icon}{Optional icon}

\item{width}{Width of the button}
}
\value{
A submit button tag
}
\description{
Create a submit button
}
