% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{suppressDependencies}
\alias{suppressDependencies}
\title{Suppress dependencies}
\usage{
suppressDependencies(...)
}
\arguments{
\item{...}{Dependency names to suppress}
}
\value{
Object marking dependencies to suppress
}
\description{
Suppress dependencies
}
