% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{tableOutput}
\alias{tableOutput}
\title{Create a table output element}
\usage{
tableOutput(outputId)
}
\arguments{
\item{outputId}{Output variable for the table}
}
\value{
A table output tag
}
\description{
Create a table output element
}
