% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-navigation.R
\name{tabPanel}
\alias{tabPanel}
\title{Create a tab panel}
\usage{
tabPanel(title, ..., value = title, icon = NULL)
}
\arguments{
\item{title}{Tab title}

\item{...}{Tab content}

\item{value}{Value used to identify tab}

\item{icon}{Optional icon}
}
\value{
A tab panel tag with metadata
}
\description{
Create a tab panel
}
