% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-navigation.R
\name{tabPanelBody}
\alias{tabPanelBody}
\title{Create a tab panel body (content only)}
\usage{
tabPanelBody(value, ...)
}
\arguments{
\item{value}{Value identifying this tab}

\item{...}{Tab content}
}
\value{
A tab panel body tag
}
\description{
Create a tab panel body (content only)
}
