% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-navigation.R
\name{tabsetPanel}
\alias{tabsetPanel}
\title{Create a tabset panel}
\usage{
tabsetPanel(
  ...,
  id = NULL,
  selected = NULL,
  type = c("tabs", "pills", "hidden"),
  vertical = FALSE,
  header = NULL,
  footer = NULL
)
}
\arguments{
\item{...}{Tab panels created with tabPanel()}

\item{id}{Optional id for the tabset (allows programmatic control)}

\item{selected}{Value of initially selected tab}

\item{type}{Visual type: "tabs", "pills", or "hidden"}

\item{vertical}{Stack tabs vertically?}

\item{header}{Header content above tabs}

\item{footer}{Footer content below tabs}
}
\value{
A tabset panel tag
}
\description{
Create a tabset panel
}
