% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tag}
\alias{tag}
\title{Create an HTML tag}
\usage{
tag(name, attribs = list(), children = list(), .noWS = NULL)
}
\arguments{
\item{name}{Tag name (e.g., "div", "span")}

\item{attribs}{Named list of attributes}

\item{children}{List of child elements}

\item{.noWS}{Character vector indicating where whitespace should be suppressed}
}
\value{
A shiny.tag object
}
\description{
Create an HTML tag
}
