% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tag_to_html}
\alias{tag_to_html}
\title{Convert tag or tag list to HTML string}
\usage{
tag_to_html(x)
}
\arguments{
\item{x}{Tag, tag list, or other content}
}
\value{
HTML string
}
\description{
Convert tag or tag list to HTML string
}
