% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tag_to_html_internal}
\alias{tag_to_html_internal}
\title{Internal function to convert tag structure to HTML}
\usage{
tag_to_html_internal(tag)
}
\arguments{
\item{tag}{Tag structure (list with name, attribs, children)}
}
\value{
HTML string
}
\description{
Internal function to convert tag structure to HTML
}
