% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagAppendAttributes}
\alias{tagAppendAttributes}
\title{Append attributes to a tag}
\usage{
tagAppendAttributes(tag, ...)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{...}{Named attributes to add}
}
\value{
Modified tag
}
\description{
Append attributes to a tag
}
