% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagAppendChild}
\alias{tagAppendChild}
\title{Append a child to a tag}
\usage{
tagAppendChild(tag, child, .cssSelector = NULL)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{child}{Child element to append}

\item{.cssSelector}{Optional CSS selector (not implemented yet)}
}
\value{
Modified tag
}
\description{
Append a child to a tag
}
