% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagAppendChildren}
\alias{tagAppendChildren}
\title{Append multiple children to a tag}
\usage{
tagAppendChildren(tag, ..., .cssSelector = NULL, list = NULL)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{...}{Children to append}

\item{.cssSelector}{Optional CSS selector (not implemented yet)}

\item{list}{Optional list of children}
}
\value{
Modified tag
}
\description{
Append multiple children to a tag
}
