% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagGetAttribute}
\alias{tagGetAttribute}
\title{Get an attribute from a tag}
\usage{
tagGetAttribute(tag, attr)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{attr}{Attribute name}
}
\value{
Attribute value or NULL
}
\description{
Get an attribute from a tag
}
