% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagHasAttribute}
\alias{tagHasAttribute}
\title{Check if a tag has an attribute}
\usage{
tagHasAttribute(tag, attr)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{attr}{Attribute name to check}
}
\value{
Logical
}
\description{
Check if a tag has an attribute
}
