% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagInsertChildren}
\alias{tagInsertChildren}
\title{Insert children at specific position}
\usage{
tagInsertChildren(tag, after, ..., .cssSelector = NULL)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{after}{Index after which to insert}

\item{...}{Children to insert}

\item{.cssSelector}{Optional CSS selector}
}
\value{
Modified tag
}
\description{
Insert children at specific position
}
