% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagList}
\alias{tagList}
\title{Combine UI elements into a list}
\usage{
tagList(...)
}
\arguments{
\item{...}{UI elements to combine}
}
\value{
A shiny.tag.list object
}
\description{
Combine UI elements into a list
}
