% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\docType{data}
\name{tags}
\alias{tags}
\title{HTML Tags Object}
\format{
An object of class \code{shiny.tags} (inherits from \code{environment}) of length 115.
}
\usage{
tags
}
\description{
An environment containing functions for all standard HTML tags.
Access tags using \code{tags$tagname}, e.g., \code{tags$div()}, \code{tags$span()}.
}
\keyword{datasets}
