% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{tagSetChildren}
\alias{tagSetChildren}
\title{Set children of a tag (replacing existing)}
\usage{
tagSetChildren(tag, ..., .cssSelector = NULL, list = NULL)
}
\arguments{
\item{tag}{A shiny.tag object}

\item{...}{New children}

\item{.cssSelector}{Optional CSS selector (not implemented yet)}

\item{list}{Optional list of children}
}
\value{
Modified tag
}
\description{
Set children of a tag (replacing existing)
}
