% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{textAreaInput}
\alias{textAreaInput}
\title{Create a textarea input control}
\usage{
textAreaInput(
  inputId,
  label,
  value = "",
  width = NULL,
  height = NULL,
  cols = NULL,
  rows = NULL,
  placeholder = NULL,
  resize = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{width}{Width of the input}

\item{height}{Height of the input}

\item{cols}{Number of columns (overrides width)}

\item{rows}{Number of rows (overrides height)}

\item{placeholder}{Placeholder text}

\item{resize}{Resize behavior: "both", "none", "vertical", "horizontal"}
}
\value{
A textarea input tag
}
\description{
Create a textarea input control
}
