% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{textInput}
\alias{textInput}
\title{Create a text input control}
\usage{
textInput(inputId, label, value = "", width = NULL, placeholder = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{width}{Width of the input (e.g., "100%", "400px")}

\item{placeholder}{Placeholder text}
}
\value{
A text input tag
}
\description{
Create a text input control
}
