% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{textOutput}
\alias{textOutput}
\title{Create a text output element}
\usage{
textOutput(outputId, container = if (inline) span else div, inline = FALSE)
}
\arguments{
\item{outputId}{Output variable to read the value from}

\item{container}{Container tag function}

\item{inline}{Use inline element (span) instead of block (div)?}
}
\value{
A text output tag
}
\description{
Create a text output element
}
