% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-time-travel.R
\name{TimeTravelDebugger}
\alias{TimeTravelDebugger}
\title{Time-Travel Debugger}
\description{
Records state history and enables replay
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeTravelDebugger-new}{\code{TimeTravelDebugger$new()}}
\item \href{#method-TimeTravelDebugger-record_snapshot}{\code{TimeTravelDebugger$record_snapshot()}}
\item \href{#method-TimeTravelDebugger-get_snapshot}{\code{TimeTravelDebugger$get_snapshot()}}
\item \href{#method-TimeTravelDebugger-get_current}{\code{TimeTravelDebugger$get_current()}}
\item \href{#method-TimeTravelDebugger-go_to_step}{\code{TimeTravelDebugger$go_to_step()}}
\item \href{#method-TimeTravelDebugger-step_forward}{\code{TimeTravelDebugger$step_forward()}}
\item \href{#method-TimeTravelDebugger-step_backward}{\code{TimeTravelDebugger$step_backward()}}
\item \href{#method-TimeTravelDebugger-replay}{\code{TimeTravelDebugger$replay()}}
\item \href{#method-TimeTravelDebugger-clear}{\code{TimeTravelDebugger$clear()}}
\item \href{#method-TimeTravelDebugger-enable}{\code{TimeTravelDebugger$enable()}}
\item \href{#method-TimeTravelDebugger-disable}{\code{TimeTravelDebugger$disable()}}
\item \href{#method-TimeTravelDebugger-get_summary}{\code{TimeTravelDebugger$get_summary()}}
\item \href{#method-TimeTravelDebugger-clone}{\code{TimeTravelDebugger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$new(max_history = 1000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-record_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-record_snapshot}{}}}
\subsection{Method \code{record_snapshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$record_snapshot(graph, state_manager, metadata = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-get_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-get_snapshot}{}}}
\subsection{Method \code{get_snapshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$get_snapshot(step)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-get_current"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-get_current}{}}}
\subsection{Method \code{get_current()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$get_current()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-go_to_step"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-go_to_step}{}}}
\subsection{Method \code{go_to_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$go_to_step(step, app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-step_forward"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-step_forward}{}}}
\subsection{Method \code{step_forward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$step_forward(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-step_backward"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-step_backward}{}}}
\subsection{Method \code{step_backward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$step_backward(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-replay"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-replay}{}}}
\subsection{Method \code{replay()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$replay(app, callback = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-clear"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-enable"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-enable}{}}}
\subsection{Method \code{enable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$enable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-disable"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-disable}{}}}
\subsection{Method \code{disable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$disable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-get_summary"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-get_summary}{}}}
\subsection{Method \code{get_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$get_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeTravelDebugger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeTravelDebugger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeTravelDebugger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
