% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{titlePanel}
\alias{titlePanel}
\title{Create a title panel}
\usage{
titlePanel(title, windowTitle = title)
}
\arguments{
\item{title}{Title text}

\item{windowTitle}{Browser window title}
}
\value{
A title panel tag
}
\description{
Create a title panel
}
