% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{uiOutput}
\alias{uiOutput}
\title{Create a UI output element}
\usage{
uiOutput(
  outputId,
  inline = FALSE,
  container = if (inline) span else div,
  fill = FALSE,
  ...
)
}
\arguments{
\item{outputId}{Output variable to read the value from}

\item{inline}{Use inline element?}

\item{container}{Container function}

\item{fill}{Whether to allow filling}

\item{...}{Additional attributes}
}
\value{
A UI output tag
}
\description{
Alias for htmlOutput for rendering UI elements.
}
