% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateActionButton}
\alias{updateActionButton}
\title{Update an action button}
\usage{
updateActionButton(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  icon = NULL,
  disabled = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{icon}{New icon}

\item{disabled}{Disabled state}
}
\value{
NULL (called for side effect)
}
\description{
Update an action button
}
