% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateActionLink}
\alias{updateActionLink}
\title{Update an action link}
\usage{
updateActionLink(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  icon = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{icon}{New icon}
}
\value{
NULL (called for side effect)
}
\description{
Update an action link
}
