% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateCheckboxInput}
\alias{updateCheckboxInput}
\title{Update a checkbox input}
\usage{
updateCheckboxInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{value}{New value (TRUE/FALSE)}
}
\value{
NULL (called for side effect)
}
\description{
Update a checkbox input
}
