% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateDateInput}
\alias{updateDateInput}
\title{Update a date input}
\usage{
updateDateInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{value}{New value}

\item{min}{New minimum date}

\item{max}{New maximum date}
}
\value{
NULL (called for side effect)
}
\description{
Update a date input
}
