% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateDateRangeInput}
\alias{updateDateRangeInput}
\title{Update a date range input}
\usage{
updateDateRangeInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  start = NULL,
  end = NULL,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{start}{New start date}

\item{end}{New end date}

\item{min}{New minimum date}

\item{max}{New maximum date}
}
\value{
NULL (called for side effect)
}
\description{
Update a date range input
}
