% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateNavbarPage}
\alias{updateNavbarPage}
\title{Update a navbar page}
\usage{
updateNavbarPage(
  session = getDefaultReactiveDomain(),
  inputId,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Navbar ID}

\item{selected}{Tab to select}
}
\value{
NULL (called for side effect)
}
\description{
Update a navbar page
}
