% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateNavlistPanel}
\alias{updateNavlistPanel}
\title{Update a navlist panel}
\usage{
updateNavlistPanel(
  session = getDefaultReactiveDomain(),
  inputId,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Navlist ID}

\item{selected}{Tab to select}
}
\value{
NULL (called for side effect)
}
\description{
Update a navlist panel
}
