% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateNumericInput}
\alias{updateNumericInput}
\title{Update a numeric input}
\usage{
updateNumericInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{value}{New value}

\item{min}{New minimum}

\item{max}{New maximum}

\item{step}{New step}
}
\value{
NULL (called for side effect)
}
\description{
Update a numeric input
}
