% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateQueryString}
\alias{updateQueryString}
\title{Update the browser's URL}
\usage{
updateQueryString(
  queryString,
  mode = c("push", "replace"),
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{queryString}{New query string}

\item{mode}{Update mode: "push" or "replace"}

\item{session}{Shiny session object}
}
\value{
NULL (called for side effect)
}
\description{
Update the browser's URL
}
