% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateRadioButtons}
\alias{updateRadioButtons}
\title{Update radio buttons}
\usage{
updateRadioButtons(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  choiceNames = NULL,
  choiceValues = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{choices}{New choices}

\item{selected}{New selected value}

\item{inline}{Display inline?}

\item{choiceNames}{Choice display names}

\item{choiceValues}{Choice values}
}
\value{
NULL (called for side effect)
}
\description{
Update radio buttons
}
