% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateSelectInput}
\alias{updateSelectInput}
\title{Update a select input}
\usage{
updateSelectInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{choices}{New choices}

\item{selected}{New selected value(s)}
}
\value{
NULL (called for side effect)
}
\description{
Update a select input
}
