% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateSelectizeInput}
\alias{updateSelectizeInput}
\title{Update a selectize input}
\usage{
updateSelectizeInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  options = NULL,
  server = FALSE
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{choices}{New choices}

\item{selected}{New selected value(s)}

\item{options}{Selectize options}

\item{server}{Use server-side selectize?}
}
\value{
NULL (called for side effect)
}
\description{
Update a selectize input
}
