% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateSliderInput}
\alias{updateSliderInput}
\title{Update a slider input}
\usage{
updateSliderInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  timeFormat = NULL,
  timezone = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{value}{New value (single or two values for range)}

\item{min}{New minimum}

\item{max}{New maximum}

\item{step}{New step}

\item{timeFormat}{Time format for date sliders}

\item{timezone}{Time zone}
}
\value{
NULL (called for side effect)
}
\description{
Update a slider input
}
