% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateTabsetPanel}
\alias{updateTabsetPanel}
\title{Update a tabset panel}
\usage{
updateTabsetPanel(
  session = getDefaultReactiveDomain(),
  inputId,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Tabset ID}

\item{selected}{Tab to select}
}
\value{
NULL (called for side effect)
}
\description{
Update a tabset panel
}
