% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateTextAreaInput}
\alias{updateTextAreaInput}
\title{Update a textarea input}
\usage{
updateTextAreaInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{value}{New value}

\item{placeholder}{New placeholder}
}
\value{
NULL (called for side effect)
}
\description{
Update a textarea input
}
