% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateTextInput}
\alias{updateTextInput}
\title{Update a text input}
\usage{
updateTextInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label (NULL to leave unchanged)}

\item{value}{New value (NULL to leave unchanged)}

\item{placeholder}{New placeholder (NULL to leave unchanged)}
}
\value{
NULL (called for side effect)
}
\description{
Update a text input
}
