% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-updates.R
\name{updateVarSelectInput}
\alias{updateVarSelectInput}
\title{Update a variable select input}
\usage{
updateVarSelectInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  data = NULL,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session object}

\item{inputId}{Input ID to update}

\item{label}{New label}

\item{data}{New data frame for variables}

\item{selected}{New selected variable(s)}
}
\value{
NULL (called for side effect)
}
\description{
Update a variable select input
}
