% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{urlModal}
\alias{urlModal}
\title{Create a modal dialog with a URL}
\usage{
urlModal(url, title = "Paste URL into browser:", subtitle = NULL)
}
\arguments{
\item{url}{URL to display}

\item{title}{Modal title}

\item{subtitle}{Optional subtitle}
}
\value{
A modal dialog tag
}
\description{
Create a modal dialog with a URL
}
