% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{validateCssUnit}
\alias{validateCssUnit}
\title{Validate CSS unit}
\usage{
validateCssUnit(x, unitType = NULL)
}
\arguments{
\item{x}{Value to validate}

\item{unitType}{Not used (for compatibility)}
}
\value{
Valid CSS string
}
\description{
Ensures a value is a valid CSS unit string.
}
