% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{varSelectInput}
\alias{varSelectInput}
\title{Create variable select input}
\usage{
varSelectInput(
  inputId,
  label,
  data,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = NULL,
  size = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{data}{Data frame to select variables from}

\item{selected}{Initially selected variable(s)}

\item{multiple}{Allow multiple selections?}

\item{selectize}{Use selectize.js?}

\item{width}{Width of the input}

\item{size}{Number of visible options}
}
\value{
A variable select input tag
}
\description{
Create variable select input
}
