% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inputs.R
\name{varSelectizeInput}
\alias{varSelectizeInput}
\title{Create variable selectize input}
\usage{
varSelectizeInput(
  inputId,
  label,
  data,
  selected = NULL,
  multiple = FALSE,
  width = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control}

\item{data}{Data frame to select variables from}

\item{selected}{Initially selected variable(s)}

\item{multiple}{Allow multiple selections?}

\item{width}{Width of the input}
}
\value{
A variable selectize input tag
}
\description{
Create variable selectize input
}
