% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-outputs.R
\name{verbatimTextOutput}
\alias{verbatimTextOutput}
\title{Create a verbatim text output element}
\usage{
verbatimTextOutput(outputId, placeholder = FALSE)
}
\arguments{
\item{outputId}{Output variable to read the value from}

\item{placeholder}{Display placeholder when empty?}
}
\value{
A verbatim text output tag
}
\description{
Displays text output in a fixed-width font, preserving whitespace.
}
