% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{verticalLayout}
\alias{verticalLayout}
\title{Create a vertical layout}
\usage{
verticalLayout(..., fluid = TRUE)
}
\arguments{
\item{...}{UI elements}

\item{fluid}{Use fluid width?}
}
\value{
A vertical layout tag
}
\description{
Stacks elements vertically.
}
