% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-time-travel.R
\name{visualize_graph}
\alias{visualize_graph}
\title{Visualize graph}
\usage{
visualize_graph(graph, format = "dot")
}
\arguments{
\item{graph}{ReactiveGraph}

\item{format}{Output format ("dot" or "json")}
}
\value{
Graph visualization
}
\description{
Visualize graph
}
