% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-websocket.R
\name{WebSocketServer}
\alias{WebSocketServer}
\title{WebSocket Server}
\description{
Manages WebSocket connections and communication
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WebSocketServer-new}{\code{WebSocketServer$new()}}
\item \href{#method-WebSocketServer-register_handler}{\code{WebSocketServer$register_handler()}}
\item \href{#method-WebSocketServer-on_open}{\code{WebSocketServer$on_open()}}
\item \href{#method-WebSocketServer-on_message}{\code{WebSocketServer$on_message()}}
\item \href{#method-WebSocketServer-on_close}{\code{WebSocketServer$on_close()}}
\item \href{#method-WebSocketServer-send_message}{\code{WebSocketServer$send_message()}}
\item \href{#method-WebSocketServer-send_graph_update}{\code{WebSocketServer$send_graph_update()}}
\item \href{#method-WebSocketServer-send_value_update}{\code{WebSocketServer$send_value_update()}}
\item \href{#method-WebSocketServer-send_dom_patch}{\code{WebSocketServer$send_dom_patch()}}
\item \href{#method-WebSocketServer-send_hot_reload}{\code{WebSocketServer$send_hot_reload()}}
\item \href{#method-WebSocketServer-send_ui_replace}{\code{WebSocketServer$send_ui_replace()}}
\item \href{#method-WebSocketServer-send_restore_inputs}{\code{WebSocketServer$send_restore_inputs()}}
\item \href{#method-WebSocketServer-send_error}{\code{WebSocketServer$send_error()}}
\item \href{#method-WebSocketServer-handle_user_input}{\code{WebSocketServer$handle_user_input()}}
\item \href{#method-WebSocketServer-broadcast}{\code{WebSocketServer$broadcast()}}
\item \href{#method-WebSocketServer-clone}{\code{WebSocketServer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-new"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$new(app)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-register_handler"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-register_handler}{}}}
\subsection{Method \code{register_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$register_handler(message_type, handler)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-on_open"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-on_open}{}}}
\subsection{Method \code{on_open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$on_open(ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-on_message"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-on_message}{}}}
\subsection{Method \code{on_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$on_message(ws, is_binary, message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-on_close"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-on_close}{}}}
\subsection{Method \code{on_close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$on_close(ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_message"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_message}{}}}
\subsection{Method \code{send_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_message(ws, type, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_graph_update"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_graph_update}{}}}
\subsection{Method \code{send_graph_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_graph_update(ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_value_update"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_value_update}{}}}
\subsection{Method \code{send_value_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_value_update(node_id, value, output_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_dom_patch"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_dom_patch}{}}}
\subsection{Method \code{send_dom_patch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_dom_patch(patch)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_hot_reload"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_hot_reload}{}}}
\subsection{Method \code{send_hot_reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_hot_reload(diff_summary)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_ui_replace"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_ui_replace}{}}}
\subsection{Method \code{send_ui_replace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_ui_replace(html)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_restore_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_restore_inputs}{}}}
\subsection{Method \code{send_restore_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_restore_inputs(input_values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-send_error"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-send_error}{}}}
\subsection{Method \code{send_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$send_error(ws, error_message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-handle_user_input"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-handle_user_input}{}}}
\subsection{Method \code{handle_user_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$handle_user_input(msg, ws)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-broadcast"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-broadcast}{}}}
\subsection{Method \code{broadcast()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$broadcast(type, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebSocketServer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WebSocketServer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocketServer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
