% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{wellPanel}
\alias{wellPanel}
\title{Create a well panel}
\usage{
wellPanel(...)
}
\arguments{
\item{...}{UI elements}
}
\value{
A well panel tag
}
\description{
A panel with a gray background and rounded corners.
}
