% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-layout.R
\name{withMathJax}
\alias{withMathJax}
\title{Load MathJax and typeset math expressions}
\usage{
withMathJax(...)
}
\arguments{
\item{...}{UI elements containing math}
}
\value{
UI with MathJax support
}
\description{
Load MathJax and typeset math expressions
}
