% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-modals.R
\name{withProgress}
\alias{withProgress}
\title{Execute code with progress indicator}
\usage{
withProgress(
  expr,
  min = 0,
  max = 1,
  value = min + (max - min) * 0.1,
  message = NULL,
  detail = NULL,
  style = "notification",
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE
)
}
\arguments{
\item{expr}{Expression to execute}

\item{min}{Minimum value}

\item{max}{Maximum value}

\item{value}{Initial value}

\item{message}{Initial message}

\item{detail}{Initial detail}

\item{style}{Style: "notification" or "old"}

\item{session}{Shiny session}

\item{env}{Environment for evaluation}

\item{quoted}{Is expr quoted?}
}
\value{
Result of expression
}
\description{
Execute code with progress indicator
}
