% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-tags.R
\name{withTags}
\alias{withTags}
\title{Evaluate expression with tags in scope}
\usage{
withTags(code)
}
\arguments{
\item{code}{Expression to evaluate}
}
\value{
Result of expression
}
\description{
Evaluate expression with tags in scope
}
