% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-websocket.R
\docType{data}
\name{WS_MESSAGE_TYPES}
\alias{WS_MESSAGE_TYPES}
\title{WebSocket Message Types}
\format{
An object of class \code{list} of length 8.
}
\usage{
WS_MESSAGE_TYPES
}
\description{
WebSocket Message Types
}
\keyword{datasets}
