% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-values.R
\name{ReactiveValues}
\alias{ReactiveValues}
\title{Reactive Values}
\description{
Container for reactive values (like Shiny's reactiveValues)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveValues-new}{\code{ReactiveValues$new()}}
\item \href{#method-ReactiveValues-get}{\code{ReactiveValues$get()}}
\item \href{#method-ReactiveValues-set}{\code{ReactiveValues$set()}}
\item \href{#method-ReactiveValues-names}{\code{ReactiveValues$names()}}
\item \href{#method-ReactiveValues-to_list}{\code{ReactiveValues$to_list()}}
\item \href{#method-ReactiveValues-clone}{\code{ReactiveValues$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValues-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValues-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValues$new(builder)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValues-get"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValues-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValues$get(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValues-set"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValues-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValues$set(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValues-names"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValues-names}{}}}
\subsection{Method \code{names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValues$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValues-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValues-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValues$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValues-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValues-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValues$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
